Create a chat server that listens to incoming conn
This chat listen to incoming connections and accepts them.
After that, it delegates the function to a handleConnection function
that receive the connection and broadcast a message to everybody
connected. It scans text from connection and broadcast it to every user
connected. Whenever the connection is interrupted, it broadcast all
users.

It also tracks the connected users in a dictionary and delete the user
corresponding key when the user closes the connection.